// src/api/model/friendRequestResult.ts
/**
 * Generated by orval v8.12.1 🍺
 * Do not edit manually.
 * DunbarHorizon API 명세서
 * DunbarHorizon 백엔드 API
 * OpenAPI spec version: 1.0.0
 */
import type { FriendRequestResultStatus } from './friendRequestResultStatus';
import type { FriendResult } from './friendResult';

export interface FriendRequestResult {
  id?: string;
  requester?: FriendResult;
  receiver?: FriendResult;
  status?: FriendRequestResultStatus;
  createdAt?: string;
}